+++ packages/nimble-blazor/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Thu, 02 Mar 2023 17:48:40 GMT and should not be manually modified.
+This log was last generated on Fri, 10 Mar 2023 22:17:50 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 11.3.3
+
+Fri, 10 Mar 2023 22:17:50 GMT
+
+### Patches
+
+- Programmatic table column width API ([ni/nimble@d38391e](https://github.com/ni/nimble/commit/d38391e69c9984de80bf7f4a41cfbcc0181e0d16))
 
 ## 11.2.7